I dunno if I have the wrong version of aclocal or what, but running:
aclocal -I m4

does not work; no aclocal.m4 file is generated that loads
m4/moo.m4. But if I move ./m4/moo.m4 to ./acinclude.m4 and just run
'aclocal' with no arguments it generates a correct aclocal.m4 file.

Well, whatever the problem is, it's irrelevant to the many tasks at
hand; I wanted to use a subdirectory called 'm4' because that's the
New Recommended Thing To Do. acinclude.m4 is the Previously
Recommended New Thing To Do and it will suffice for the time being.